extract_audio_function/requirements.txt
google-cloud-storage 